+++ src_3_0_0/modules/menu/components/MenuContainerBodyElement.vue
@@ -52,7 +52,6 @@ export default {
         }
     },
     mounted () {
-        console.log("MenuContainerBodyElement path", this.path);
         if (this.properties?.active) {
             this.clickedMenuElement(this.path);
         }